refactor: move comlink helper to utils
